Generate FPL H2H standings statically
The Python script in the `./scripts` directory is meant to be invoked by
the automation to statically generate the H2H standings per quarter as a
YAML file. This YAML file gets consumed via a shortcode to generate the
H2H quarter standings.

We will set up a GitHub action to regenerate the standings every day.