Copy the repository into the container during build
For GitHub Action type workflow we'll create a copy of the repository
inside the image. For development we typically want to mount the
respoitory from the host into the container so that any changes made on
the host are reflected instantly inside the container. The volume mount
takes precedence over the copy.