Create GitHub Pages and FPL H2H standings actions
The Dockerfile in the root of the repository is used to build an image
which then gets used by the GitHub Actions to generate the static
website into a gh-pages branch which is deplyed via GitHub Pages. The
same docker image is used to invoke the FPL H2H standings script on a
schedule basis to generate the data in the master branch. Upon pushing
to the master branch the Hugo build will get triggered.

This action will eventually run the build commands within the docker
container and deploy the website to a GitHub Pages branch.